fix: bloquear edição e remoção de itens emprestados

+++ acaoBotao.php
@@ -20,19 +20,29 @@ if (isset($_POST['postar'])) {
   }
 }
 if (isset($_POST['deletar'])) {
-  $id_item = $_POST['id_item'];
+  $status = $_POST['status'];
+  if ($status == "emprestado") {
+    header("Location: dashboard.php");
+  } else {
+    $id_item = $_POST['id_item'];
 
-  $query = "DELETE FROM itens
+    $query = "DELETE FROM itens
             WHERE id_item = ?";
 
-  $stmt = $conn->prepare($query);
-  $stmt->bind_param("i", $id_item);
-  $stmt->execute();
-  $stmt->close();
+    $stmt = $conn->prepare($query);
+    $stmt->bind_param("i", $id_item);
+    $stmt->execute();
+    $stmt->close();
 
-  header("Location: dashboard.php");
+    header("Location: dashboard.php");
+  }
 }
 if (isset($_POST['editar'])) {
-  $id_item = $_POST['id_item'];
-  header("Location: editarItem.php?id_item=" . $id_item);
+  $status = $_POST['status'];
+  if ($status == "emprestado") {
+    header("Location: dashboard.php");
+  } else {
+    $id_item = $_POST['id_item'];
+    header("Location: editarItem.php?id_item=" . $id_item);
+  }
 }
